make starting up slightly faster
